+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Canary
 *   Show number of unread notes and messages in the bottom right corner
-*   Fixed max ad length for automated ads to 50,000 characters
+*   Fixed max ad length for automated ads
 *   Fixed 'unsure' sexual orientation to display correctly in character preview